+++ drivers/video/imx335.c
@@ -16,8 +16,6 @@
 #include <zephyr/video/video.h>
 
 #include "video_common.h"
-#include "video_ctrls.h"
-#include "video_device.h"
 
 LOG_MODULE_REGISTER(video_imx335, CONFIG_VIDEO_LOG_LEVEL);
