+++ drivers/video/mt9m114.c
@@ -15,8 +15,7 @@
 #include <zephyr/sys/byteorder.h>
 #include <zephyr/video/video.h>
 
-#include "video_ctrls.h"
-#include "video_device.h"
+#include "video_common.h"
 
 LOG_MODULE_REGISTER(video_mt9m114, CONFIG_VIDEO_LOG_LEVEL);
