+++ drivers/video/ov5642.c
@@ -24,8 +24,6 @@
 #include <zephyr/video/video.h>
 
 #include "video_common.h"
-#include "video_ctrls.h"
-#include "video_device.h"
 
 LOG_MODULE_REGISTER(video_ov5642, CONFIG_VIDEO_LOG_LEVEL);
