+++ drivers/video/ov767x.c
@@ -14,8 +14,6 @@
 #include <zephyr/video/video.h>
 
 #include "video_common.h"
-#include "video_ctrls.h"
-#include "video_device.h"
 
 LOG_MODULE_REGISTER(video_ov767x, CONFIG_VIDEO_LOG_LEVEL);
