+++ drivers/video/video_emul_imager.c
@@ -17,8 +17,7 @@
 #include <zephyr/sys/util.h>
 #include <zephyr/video/video.h>
 
-#include "video_ctrls.h"
-#include "video_device.h"
+#include "video_common.h"
 
 LOG_MODULE_REGISTER(video_emul_imager, CONFIG_VIDEO_LOG_LEVEL);
