+++ drivers/video/video_emul_rx.c
@@ -16,7 +16,7 @@
 #include <zephyr/sys/util.h>
 #include <zephyr/video/video.h>
 
-#include "video_device.h"
+#include "video_common.h"
 
 LOG_MODULE_REGISTER(video_emul_rx, CONFIG_VIDEO_LOG_LEVEL);
