+++ drivers/video/video_stm32_dcmi.c
@@ -21,7 +21,7 @@
 
 #include <stm32_ll_dma.h>
 
-#include "video_device.h"
+#include "video_common.h"
 
 LOG_MODULE_REGISTER(video_stm32_dcmi, CONFIG_VIDEO_LOG_LEVEL);
