+++ drivers/video/video_sw_generator.c
@@ -16,8 +16,7 @@
 #include <zephyr/sys/util.h>
 #include <zephyr/video/video.h>
 
-#include "video_ctrls.h"
-#include "video_device.h"
+#include "video_common.h"
 
 LOG_MODULE_REGISTER(video_sw_generator, CONFIG_VIDEO_LOG_LEVEL);
