+++ subsys/usb/host/class/usbh_uvc.c
@@ -15,7 +15,7 @@
 #include <zephyr/usb/usb_ch9.h>
 #include <zephyr/drivers/usb/udc.h>
 #include <zephyr/drivers/video.h>
-#include <zephyr/video/controls.h>
+#include <zephyr/video/video.h>
 #include <zephyr/logging/log.h>
 #include <zephyr/sys/util.h>
 
@@ -25,8 +25,7 @@
 #include "usbh_device.h"
 
 #include "uvc.h"
-#include "../../../drivers/video/video_ctrls.h"
-#include "../../../drivers/video/video_device.h"
+#include "../../../../drivers/video/video_common.h"
 
 LOG_MODULE_REGISTER(usbh_uvc, CONFIG_USBH_UVC_LOG_LEVEL);
