+++ subsys/video/frmival.c
@@ -35,3 +35,83 @@ int video_enum_frmival(const struct device *dev, struct video_frmival_enum *fie)
 
 	return video_driver_enum_frmival(dev, fie);
 }
+
+int video_closest_frmival_stepwise(const struct video_frmival_stepwise *stepwise,
+				   const struct video_frmival *desired,
+				   struct video_frmival *match)
+{
+	if (stepwise == NULL || desired == NULL || match == NULL) {
+		return -EINVAL;
+	}
+
+	uint64_t min = stepwise->min.numerator;
+	uint64_t max = stepwise->max.numerator;
+	uint64_t step = stepwise->step.numerator;
+	uint64_t goal = desired->numerator;
+
+	/* Set a common denominator to all values */
+	min *= stepwise->max.denominator * stepwise->step.denominator * desired->denominator;
+	max *= stepwise->min.denominator * stepwise->step.denominator * desired->denominator;
+	step *= stepwise->min.denominator * stepwise->max.denominator * desired->denominator;
+	goal *= stepwise->min.denominator * stepwise->max.denominator * stepwise->step.denominator;
+
+	__ASSERT_NO_MSG(step != 0U);
+	/* Prevent division by zero */
+	if (step == 0U) {
+		return -ERANGE;
+	}
+	/* Saturate the desired value to the min/max supported */
+	goal = CLAMP(goal, min, max);
+
+	/* Compute a numerator and denominator */
+	match->numerator = min + DIV_ROUND_CLOSEST(goal - min, step) * step;
+	match->denominator = stepwise->min.denominator * stepwise->max.denominator *
+			     stepwise->step.denominator * desired->denominator;
+
+	return 0;
+}
+
+int video_closest_frmival(const struct device *dev, struct video_frmival_enum *match)
+{
+	if (dev == NULL || match == NULL || match->type == VIDEO_FRMIVAL_TYPE_STEPWISE) {
+		return -EINVAL;
+	}
+
+	struct video_frmival desired = match->discrete;
+	struct video_frmival_enum fie = {.format = match->format};
+	uint64_t best_diff_nsec = INT32_MAX;
+	uint64_t goal_nsec = video_frmival_nsec(&desired);
+
+	for (fie.index = 0; video_enum_frmival(dev, &fie) == 0; fie.index++) {
+		struct video_frmival tmp = {0};
+		uint64_t diff_nsec = 0;
+		uint64_t tmp_nsec;
+
+		switch (fie.type) {
+		case VIDEO_FRMIVAL_TYPE_DISCRETE:
+			tmp = fie.discrete;
+			break;
+		case VIDEO_FRMIVAL_TYPE_STEPWISE:
+			video_closest_frmival_stepwise(&fie.stepwise, &desired, &tmp);
+			break;
+		default:
+			CODE_UNREACHABLE;
+		}
+
+		tmp_nsec = video_frmival_nsec(&tmp);
+		diff_nsec = tmp_nsec > goal_nsec ? tmp_nsec - goal_nsec : goal_nsec - tmp_nsec;
+
+		if (diff_nsec < best_diff_nsec) {
+			best_diff_nsec = diff_nsec;
+			match->index = fie.index;
+			match->discrete = tmp;
+		}
+
+		if (diff_nsec == 0) {
+			/* Exact match, stop searching a better match */
+			break;
+		}
+	}
+
+	return 0;
+}
